+++ src/parser.rs
@@ -4,7 +4,6 @@ use nom::{branch::alt,
           character::complete::{char},
           IResult, AsChar};
 use nom::character::is_alphanumeric;
-use std::collections::HashMap;
 
 #[derive(Debug)]
 pub enum Atom {
@@ -15,29 +14,10 @@ pub enum Atom {
 }
 
 #[derive(Debug)]
-pub enum EvaluationError {
-    DivideByZero,
-    FunctionNotFound,
-    NotAFunction,
-    WrongType(String),
-    WrongArity(i64, i64) // (Expected, Received)
-}
-
-#[derive(Debug)]
 pub enum Expression {
     At(Atom),
     Expr(Box<Expression>, Vec<Expression>),
 }
-
-/* TODO: Discover how to parametrize the Ok return value
-   E.g.:  When evaluation of summation goes OK, the result is not
-   just any Expression, it is in fact an Expression that contains an integer.
-   Any value of that return Expression is going to successfully pattern match on
-   the pattern
-   Expression::At(Atom::Int(val))
-
-   How do we get this guarantee on the type level? */
-pub type EvalResult = Result<Expression, EvaluationError>;
 
 fn is_space_lisp(c: u8) -> bool {
     is_space(c) || c.as_char() == ','
@@ -138,45 +118,4 @@ pub fn parse(input: &str) -> IResult<&[u8], Expression, (&[u8], nom::error::Erro
     let expression = parse_expression_top_level(input.as_bytes());
     println!("{:?}", expression);
     expression
-}
-
-fn map_m(maybes: Vec<EvalResult>) -> Result<Vec<Expression>, EvalResult> {
-    let mut to_return = Vec::new();
-    for eval_res in maybes {
-        match eval_res {
-            Err(_) =>{ return Err(eval_res) }
-            Ok(expr) => { to_return.push(expr) }
-        }
-    }
-    Ok(to_return)
-}
-
-fn eval_expr3(expr: Expression, vars: &HashMap<String, fn(Vec<Expression>) -> EvalResult>) -> EvalResult {
-    match expr {
-        Expression::At(_) => { Ok(expr) }
-        Expression::Expr(op, args) => {
-            let evaled_fn_symbol = eval_expr3(*op, vars)?;
-            let evaled_args = args.into_iter().map(|x| eval_expr3(x, vars));
-            let try_correct_evaled_args = map_m(evaled_args.collect());
-            match try_correct_evaled_args {
-                Err(err) => { err }
-                Ok(evaled_args) => {
-                    match evaled_fn_symbol {
-                        Expression::At(Atom::Symbol(sym)) => {
-                            match vars.get(&sym) {
-                                None => { Err(EvaluationError::FunctionNotFound) }
-                                Some(rust_fn) => { rust_fn(evaled_args) }
-                            }
-                        }
-                        _ => {Err(EvaluationError::NotAFunction)}
-                    }
-                }
-            }
-        }
-    }
-}
-
-pub fn eval(expr: Expression, vars: &HashMap<String, fn(Vec<Expression>) -> EvalResult>) -> String {
-    let evaled_expr = eval_expr3(expr, vars);
-    format!("{:?}", evaled_expr)
 }
